requirements.txt
fastapi
fastapi[standard]

huggingface_hub

Pillow